requirements.txt
flask
dnspython
requests
python-whois
